AES wrap/unwrap implement

+++ app/nand/nand/develop/include/DeviceInfo.h
@@ -34,7 +34,7 @@ struct NandInfo
     uint16_t bbt_format;            // 1
     string   boot_from_nand;        // AQAAA==
     uint64_t device_readid;         // 848619416
-    int64_t  dumped_page_size; 
+    uint32_t dumped_page_size;
     uint16_t meta_per_logical_page; // 12
     bool     metadata_whitening;    // "AQAAA==" => "01000000" => true
     string   name;
